Rename, Remove: frontend-setting
