--- client/src/App.js
@@ -1,24 +1,0 @@
-import React from "react";
-import "./App.css";
-import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import Navbar from "./components/Navbar/Navbar";
-import Home from "./components/Home/Home";
-import Register from "./components/Register/Register";
-import Login from "./components/Login/Login";
-
-function App() {
-  return (
-    <>
-      <Router>
-        <Navbar />
-        <Routes>
-          <Route path="/" exact element={<Home />}></Route>
-          <Route path="/register" exact element={<Register />}></Route>
-          <Route path="/login" exact element={<Login />}></Route>
-        </Routes>
-      </Router >
-    </>
-  );
-}
-
-export default App;
